EGL context protection extension and use of it in the HWUI library
This change is a diagnostic feature for logging some OpenGL calls
made by applications on the HWUI EGLContext. This is done with an
EGL extension, eglProtectContextSEC, that can be used to change
the protected state of the given EGLContext. When protected, some
OpenGL calls on the context are logged. The HWUI EGLContext is
protected after creation and is unprotected in the HWUI library
before calling the intercepted OpenGL calls and then re-protected
after the calls. Changing the protection state only involves changing
a function table pointer so these new protect/unprotect calls in the
HWUI library are very lightweight. Unprotected contexts see no
overhead at all with this change.

Change-Id: I3fffc353a89fc6913b31ffac17032d63cb30537c
Signed-off-by: Raj Mamadgi <rmamadgi@sta.samsung.com>