installd movefiles, preserve permission, selinux labels on directories
The installd movefiles command can be used to move files from one
package to another, for instance in case of a change in package name.
During the move process, files are moved with rename() and directories
are recreated with mkdir(). This rename and mkdir process results in
the loss of permissions and selinux labels. Upodated applications can
be rendered completely unusable.

Unfortunately it's not possible to run restorecon on newly created
directories since, /data/system/packages.list may not be updated yet
with new package.

Change-Id: Id260a7af934ee382cf8b2db5bc26fc0c7970f7bd